refactor(PeoplePicker): move all people picker files into single sub-folder of 'ui' it's far easier to develop this feature if all the related files are sitting next to each other and it will be easier to add this into the staff dashboard if the Control & Modal aren't sitting under EditorsPage in the SubmissionWizard
