+++ app/components/ui/PeoplePicker/SearchBox.js
