+++ app/components/ui/PeoplePicker/SearchBox.test.js
